+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Gpio: (Breaking change) Remove InterruptError. Merge remaining errors with Error.
 * Gpio: (Breaking change) Replace all DevGpioMem and DevMem errors with Error::PermissionDenied and Error::Io.
 * Gpio: (Breaking change) Change the return value for poll_interrupt() and poll_interrupts() to Ok(Option) on success, with Some() indicating an interrupt triggered, and None indicating a timeout occurred.
+* Gpio: (Breaking change) Only a single instance of Gpio can exist at any time. Creating another instance before the existing one goes out of scope will return an Error::InstanceExists.
 * Spi: Add support for SPI with half-duplex reads/writes and full-duplex multi-segment transfers.
 
 ## 0.5.1 (May 19, 2018)